work/GoodLuck/server/app/requirements.txt
absl-py==2.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
billiard==4.2.4
blinker==1.9.0
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==41.0.7
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.1
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flatbuffers==25.9.23
fonttools==4.61.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
jiter==0.12.0
kiwisolver==1.4.9
kombu==5.6.1
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.21
ml_dtypes==0.5.4
moviepy==2.2.1
numpy==1.26.4
openai==2.8.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
proglog==0.1.12
prompt_toolkit==3.0.52
protobuf==4.25.8
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
scipy==1.16.3
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
